refactoring: move methods into a validator class
